+++ dd_yandex_pay/yp_client.py
@@ -7,12 +7,8 @@ from typing import Union
 
 import requests
 
-
-PRODUCTION_URL: Final[str] = "https://pay.yandex.ru/api/merchant/"
-"""Боевой сервер"""
-
-SANDBOX_URL: Final[str] = "https://sandbox.pay.yandex.ru/api/merchant/"
-"""Тестовая среда"""
+from dd_yandex_pay.constants import BASE_URL_PRODUCTION
+from dd_yandex_pay.exceptions import YandexPayAPIError
 
 
 class YandexPayClient:
@@ -22,12 +18,13 @@ class YandexPayClient:
     [yandex_pay_api_docs]: https://pay.yandex.ru/ru/docs/custom/backend/yandex-pay-api/
     """
 
-    RESOURCE_ORDER = "v1/orders"
+    RESOURCE_ORDER_CREATE: Final[str] = "v1/orders"
+    RESOURCE_ORDER_DETAILS: Final[str] = "v1/orders/{id}"
 
     def __init__(
         self,
         api_key: str,
-        base_url: Optional[str] = PRODUCTION_URL,
+        base_url: Optional[str] = BASE_URL_PRODUCTION,
         timeout: Optional[Union[float, int, tuple]] = (3, 10),
         deadline: Optional[int] = 10 * 1000,
     ):
@@ -38,6 +35,7 @@ class YandexPayClient:
             timeout: Дефолтный таймаут запроса для [requests](https://requests.readthedocs.io/en/latest/user/advanced/#timeouts).
             deadline: Дефолтный таймаут запроса в миллисекундах для [яндекс](https://pay.yandex.ru/ru/docs/custom/backend/yandex-pay-api/).
         """
+
         self.api_key = api_key
         self.base_url = base_url
         self.default_timeout = timeout
@@ -45,7 +43,7 @@ class YandexPayClient:
 
     # Helpers
 
-    def get_headers(self, custom: Optional[dict] = None) -> dict:
+    def get_headers(self, **custom: dict) -> dict:
         """
         Создаёт заголовки для запроса.
 
@@ -62,7 +60,7 @@ class YandexPayClient:
             "X-Request-Id": str(uuid.uuid4()),
             "X-Request-Timeout": str(self.default_deadline),
             # "X-Request-Attempt": 1, ???
-            **(custom or {}),
+            **custom,
         }
 
     def get_url(self, resourse: str) -> str:
@@ -83,7 +81,6 @@ class YandexPayClient:
         method: str,
         url: str,
         headers: Optional[dict] = None,
-        raise_errors: Optional[bool] = True,
         **kwargs: dict,
     ) -> requests.Response:
         """
@@ -93,23 +90,52 @@ class YandexPayClient:
             method: Метод запроса.
             url: Запрашиваемый ресурс.
             headers: Объект с заголовками.
-            raise_errors: Объект с кастомными заголовками.
             kwargs: Объект с кастомными заголовками.
 
         Returns:
             Полученый ответ.
         """
+
         kwargs.setdefault("timeout", self.default_timeout)
-        kwargs["headers"] = self.get_headers(headers)
+        kwargs["headers"] = self.get_headers(**(headers or {}))
         response = requests.request(method, url, **kwargs)
-
-        if raise_errors:
-            """
-            Проверять наличие ошибки при 200 статусе?
-            """
-            response.raise_for_status()
-
         return response
+
+    def response_handler(
+        self,
+        response: requests.Response,
+        check_data_availability: Optional[bool] = False,
+    ) -> dict:
+        """
+        Обработчик данных ответа.
+
+        Attributes:
+            response: Ответ API Yandex Pay.
+            check_data_availability: Флаг, указывающий на необходимость проверки на наличие
+                параметра `data` в ответе.
+
+        Returns:
+            Данные ответа API.
+
+        Raises:
+            requests.exceptions.HTTPError: HTTP Errors.
+            dd_yandex_pay.exceptions.YandexPayAPIError: API Errors.
+        """
+
+        response.raise_for_status()
+        response_data = response.json()
+
+        if response_data.get("status", "") != "success":
+            msg = response_data.get("reason", "Unknown error")
+            code = response_data.get("reasonCode", "unknown_error")
+            raise YandexPayAPIError(response, msg, code)
+
+        if check_data_availability and "data" not in response_data:
+            msg = "Response has no data."
+            code = "has_no_data"
+            raise YandexPayAPIError(response, msg, code)
+
+        return response_data
 
     # API
 
@@ -122,11 +148,10 @@ class YandexPayClient:
         availablePaymentMethods: Optional[List[str]] = None,
         extensions: Optional[dict] = None,
         ttl: Optional[int] = None,
-        raise_errors: Optional[bool] = True,
         **kwargs: dict,
-    ) -> requests.Response:
+    ) -> dict:
         """
-        Запрос для создания ссылки на оплату.
+        Запрос на создание ссылки на оплату заказа.
 
         Подбронее о данных и ответе в документации [яндекса](https://pay.yandex.ru/ru/docs/custom/backend/yandex-pay-api/order/merchant_v1_orders-post).
 
@@ -140,22 +165,22 @@ class YandexPayClient:
             extensions: [Дополнительные параметры](https://pay.yandex.ru/ru/docs/custom/backend/yandex-pay-api/order/merchant_v1_orders-post#orderextensions)
                 для оформления оффлайн заказа.
             ttl: Время жизни заказа (в секундах).
-            raise_errors: Флаг, вызывать ошибки с помощью метода
-                [raise_for_status][requests.Response.raise_for_status] или нет.
             kwargs: Прочие дополнительные параметры метода [request][requests.request] кроме method,
                 url и json.
 
         Returns:
-            Полученый ответ API.
+            Данные ответа на создание ссылки для оплаты.
+
+        Raises:
+            requests.exceptions.HTTPError: HTTP Errors.
+            dd_yandex_pay.exceptions.YandexPayAPIError: API Errors.
         """
+
         json = {
-            # "availablePaymentMethods": availablePaymentMethods,
             "cart": cart,
             "currencyCode": currencyCode,
-            # "extensions": extensions,
             "orderId": orderId,
             "redirectUrls": redirectUrls,
-            # "ttl": ttl,
         }
 
         if availablePaymentMethods:
@@ -169,10 +194,38 @@ class YandexPayClient:
 
         response = self.request(
             "POST",
-            self.get_url(self.RESOURCE_ORDER),
+            self.get_url(self.RESOURCE_ORDER_CREATE),
             json=json,
-            raise_errors=raise_errors,
             **kwargs,
         )
 
-        return response
+        response_data = self.response_handler(response, True)
+        return response_data["data"]
+
+    def get_order(self, order_id: str, **kwargs: dict) -> dict:
+        """
+        Запрос на получение деталей заказа.
+
+        Подбронее о данных и ответе в документации [яндекса](https://pay.yandex.ru/ru/docs/custom/backend/yandex-pay-api/order/merchant_v1_order-get).
+
+        Attributes:
+            order_id: Идентификатор заказа.
+            kwargs: Прочие дополнительные параметры метода [request][requests.request] кроме method,
+                url и json.
+
+        Returns:
+            Полученные данные заказа.
+
+        Raises:
+            requests.exceptions.HTTPError: HTTP Errors.
+            dd_yandex_pay.exceptions.YandexPayAPIError: API Errors.
+        """
+
+        response = self.request(
+            "GET",
+            self.get_url(self.RESOURCE_ORDER_DETAILS.format(id=order_id)),
+            **kwargs,
+        )
+
+        response_data = self.response_handler(response, True)
+        return response_data["data"]